Fix annoyance in Chrome on small screens
Previously, opening the dropdown and closing it with the hamburger
button would scroll down below the header bar for a pretty stupid
reason: the layer that greys out the page while the dropdown is open is
actually the completely re-styled chceckbox that the hamburger button is
the label to and Chrome helpfully scrolls to a button when it isn't in
view and its label is clicked.

We fix this by making the overlay fill the viewport, but now we have to
adjust the header bar to be rendered in front of the overlay. As it
turns out, this is not as straight-forward as having just the dropdown
in front of the overlay as was the case before, because the dropdown
isn't a parent element of the dropdown checkbox, while the header bar
is (there is a way around this involving a negative z-index, but going
that route renders the overlay un-clickable; I didn't look into why
exactly that is the case though). The solution I found is moving all
parts of the DOM that should be in front of the overlay out of the
direct parent path of the overlay checkbox, and then setting their
z-index higher than the overlay's z-index.