Rewrite site-nav dropdown page overlay
This mostly reverts e7825a4. The overlay has the same positioning as
before again, but now the overlay is not the restyled checkbox anymore,
but rather a secondary label for it. Because of that, the actual
checkbox still has `display: none` even when the dropdown is open, and
so Chrome won't attempt to scroll down when the dropdown is closed
again. This was necessary because Firefox on Android didn't fully apply
all of the style changes to the checkbox, and still rendered the check
mark. It also simplifies the CSS though.